--- forGui/root.py
@@ -1,47 +1,0 @@
-import tkinter as tk
-
-
-class RollOutMethodInterface(tk.Tk):
-    def __init__(self, *args, **kwargs):
-        tk.Tk.__init__(self, *args, **kwargs)
-
-        container = tk.Frame(self)
-        container.pack(side="top", fill="both", expand=True)
-
-        self.frames = {}
-        for F in (HomePage, CalibrationPage):
-            frame = F(container, self)
-            self.frames[F] = frame
-            frame.grid(row=0, column=0, sticky="nsew")
-
-        self.show_frame(HomePage)
-
-    def show_frame(self, cont):
-        frame = self.frames[cont]
-        frame.tkraise()
-
-
-class HomePage(tk.Frame):
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-        self.controller = controller
-        self.create_widgets()
-
-    def create_widgets(self):
-        self.title_label = tk.Label(self, text="Home Page")
-        self.title_label.pack()
-
-
-class CalibrationPage(tk.Frame):
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-        self.controller = controller
-        self.create_widgets()
-
-    def create_widgets(self):
-        self.title_label = tk.Label(self, text="Calibration Page")
-        self.title_label.pack()
-
-
-appROM = RollOutMethodInterface()
-appROM.mainloop()
